third_party/blink/web_tests/platform/generic/fast/selectors/query-update-distribution-expected.txt

Should update distribution when needed for querySelector and related methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS aRoot.querySelector(':host-context(#c) #d') is d
PASS toArray(aRoot.querySelectorAll(':host-context(#c) #d')) is [d]
PASS hostRoot.querySelector('::slotted(#a)') is null
PASS toArray(hostRoot.querySelectorAll('::slotted(#a)')) is []
PASS a.matches('::slotted(#a)') is false
PASS d.matches(':host-context(#host) #d') is true
PASS d.matches(':host-context(#c) #d') is true
PASS b.closest('::slotted(#a)') is null
PASS e.closest(':host-context(#host) #d') is d
PASS successfullyParsed is true

TEST COMPLETE
